Tell tox to use develop mode
We must do this, because py.test adjusts sys.path to make
'tests.test_foo' importable [*].  When py.test does this, the
'linkcheck' directory at the top of the git tree is the one that gets
imported in the tests.  If we've told pip to use develop mode, all's
fine.  If we haven't, then we're going to get errors because extension
modules like _network.so get installed into
.tox/*/lib/*/site-packages/linkcheck/network and not into
./linkcheck/network/

[*] http://doc.pytest.org/en/latest/goodpractices.html#choosing-a-test-layout-import-rules