use more consistently "maintainer" instead of "member"
tough call, but i think it actually matters... In the code of conduct,
a "community member" is not necessarily an official status. by
conflating the "maintainer" and "member" status, we would have
diminished the role of non-official members... People are a member of
the community when they participate, regardless of their official
status, and deserve the same respect as anyone else.

also use "administrator" instead of "owner". i do not "own" this
project, i administer it. we do not have to comply with GitHub's
proprietary vocabulary