introduce demotion rules
While reviewing our contribution guidelines in #139, we found we had no formal way of removing disappeared or idle maintainers, at least not explicitly, in the Membership section. This was a problem to get changes done on the contributing document which requires consensus from the maintainers.

This change introduces a way to make changes to the document by using it as a "vote of confidence", in a way: if a change is proposed and an administrator does not respond within a given 1-month delay, the administrator can be demoted down to a regular maintainer and the document approved.

Closes: #139