Begin to resurrent AnchorCheck
I started with a test of urlencoded anchors, assuming at the URL might
have a urlencoded anchor, but the actual anchor in the HTML would NOT be
urlencoded.

Then I spent a long time tearing my hair out trying to understand why the
anchors were disappearing from the variables, before realizing that
fileurl.py was stripping them. So I stopped it from doing that, and
tried to fix up all the places that were assuming the url would map to a
filesystem file. The tests all pass, but I'm not 100% sure I caught all
the cases, or fixed them correctly.

A part of that was because I was seeing inconsistency between
urlbase.py's `anchor` variable and `url_parts[4]`. I think that was just
because of fileurl.py - but it seems fraught with peril to store the
same information in two places (i.e. they get out of sync!). I think
I've fixed it for anchors, but I'm pretty sure `query` still has this
problem. (I'm not going to try to fix that.)

Some of my new tests are showing results that I'm pretty sure are
incomplete, but I'm committing now because everything is in a consistent
state. I've also identified a few places where I know there is still
wrong behavior (and need test cases to demonstrate it); I've annotated
those with 'XXX'. I'm also still seeing incomplete results in a manual
site scan. So there's more to do!