Fix a major throughput bottleneck for HTTP URLs
From my reading of the code, it looks like Aggregator / HttpUrl are
hard-coded to throttle requests to an HTTP server to a max of (an
average of) four times a second, even though the default documented
maxrequestspersecond is 10. This is because of code that was introduced
9 years ago (!) in commit 7b34be590, such that the wait_for_host method
will always wait an average of at least a quarter of a second.

There's no explanation in the commit for why this was done, and no
reasonable explanation that I can understand for it. So I'm just taking
it out.

FWIW, this made runs that were running incredibly slowly against my local
HTTP server (i.e. minutes for HTTP vs 6 seconds for local file access)
run comparably-fast over both HTTP and local file access, after turning
maxrequestspersecond way up.

Note that I can still get back to the old slow performance, almost
exactly, by setting maxrequestspersecond to 4 :)

Note that this is unrelated to the rest of the commits on this branch,
but I noticed it once I started doing manual performance testing, and I
realized it could be easily fixed, so I'm doing so in a separate commit.