Add lots of caching to improve AnchorCheck performance
Also removed the documentation saying AnchorCheck is disabled.

A downside to the prior AnchorCheck implementation was that it caused the
system to completely re-process a given URL (i.e. a document) each time
there was a URL that referenced a new (distinct) anchor in that document.
That caused enormous slowdowns. This commit introduces a new cache class
and three new places that use that cache, which avoids most of the
reprocessing and massively improves performance.

Note, however, that even under this implementataion there is still
substantial wasted time in e.g. opening new HTTP connections to servers
and redoing processing that has already been done once. I don't see any
way to remediate that without major invasive surgery.

For future reference, I did manual performance testing, with and
without this change, on a test set of about 150 reasonably-large HTML
files generated from markdown documentation (mkdocs).

These times are based on the performance improvements I made in commit
c3e97b96.

The basic config was:

[checking]
threads=0
maxrequestspersecond=10000
[filtering]
checkextern=0
[output]
fileoutput=html
log=none

Without [AnchorCheck], the times to check 478 links were:

local files: 4 seconds
local dev webserver: 6 seconds

With [AnchorCheck] and anchorcachesize=0 (which effectively disables
this cache), the times to check 1753 links (each link with a distinct
anchor counts, now) were:

local files: 57 seconds
local dev webserver: 68 seconds

With [AnchorCheck] and anchorcachesize=2000 (the default, and big enough
for my entire dataset), this cache), the times to check 1753 links were:

local files: 5 seconds
local dev webserver: 16 seconds
(changing threads to 5 reduces the time to 11 seconds)

(Without commit b4455d51, but with 5 threads, against the local dev
webserver, the time would have been 390 seconds, even with this cache!)

I have not tested against a remote / high-latency / slow web server; I
expect anchor checking might still be substantially slower than
non-anchor-checking in that case.

Note that there is some overhead introduced here in the non-AnchorCheck
case, for calling the synchorinzed get() and put() methods on the cache.
That didn't seem to make a meaningful difference in my testing, but it's
possible there are scenarios where it could matter. If so, the
synchronization to be moved to private methods and the public methods
could check the cache size to see if it's 0 before trying to use it.

Note that there is still probably documentation work to be done for this
new setting (anchorcachesize) - I updated the places I could understand
to update, but I don't think I got everywhere.