Fix a major thread-safety bug in AnchorCheck
Also stop using UrlBase.anchors in favor of a new method that is based
on UrlBase.url_data[4], which is more consistent with the rest of the
code.

The threading issue has been there for years, but I didn't notice it
until after I thought I was done, while I was doing manual testing
(with threads re-enabled).

The problem was with storing URL-specific state (.anchors) on the
AnchorCheck object itself, because there's only one global AnchorCheck
object, so all the threads are competing to use that one simgle variable
(self.anchors).

The solution was to create a new object to hold .anchors, for each
processed URL.