Add a feature to report all references to a URL
Often when there is a broken URL, e.g. from documentation changes that
moved a file or changed an anchor, there are many other pages that
reference that URL. Before this commit, linkchecker would only report
the first reference to that broken URL. With this commit, linkchecker is
now able to optionally report all references to that broken url.

Command line argument: --allrefs

Config setting: output.reportallreferences

It is off by default, mainly because I didn't want to risk breaking
existing behavior people might expect, and it would take a lot of
existing-test overhaul. There is very little performance penalty to
enabling it.

I intentionally built this on top of my other AnchorCheck code, because
I need it to work with anchor checks.